[SYCL] Add a leaf limit to the execution graph
This patch adds a leaf limit (per memory object) for the command
execution graph in order to avoid leaf bloat in applications that have
an overwhelming number of command groups that can be executed in
parallel. Limiting the number of leaves is necessary for reducing
performance overhead of regular cleanup of finished command nodes.

Whenever the limit is exceeded, the oldest leaf is added as a
dependency of the new one instead.

Signed-off-by: Sergey Semenov <sergey.semenov@intel.com>