[SYCL][Driver] Improve fat static library support
When processing libraries on the command line, take into account any thusly
named static archive (*.a) file and consider that for offloading.  We will
also scan the appropriate linker options passed in so we can determine if
the library should be considered as whole-archive.  The static libraries
found on the command line will be 'sniffed' to determine if the static
library is fat.

This effectively negates the need to use -foffload-static-lib and
-foffload-whole-static-lib which we should consider deprecated now.

Add a deprecated diagnostic when -foffload-static-lib is used

Refactor to bring along common code for the device check.  Narrows the focus
of what is considered to be processed from the linker only arguments.  Pulls
in objects from -Wl, instead of only archives, but is only part of a potential
partial link step and is not fully processed.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>