[Driver][SYCL] Improve dependency file behaviors for -fintelfpga (#1186)
When compiling AOT for FPGA dependency files are generated that are used
by the aoc compilation.  Single step compilation is seamless as the dependency
file is generated then immediately used.  When compiling to object, keeping
track of the dependency file is not so intuitive.

To alleviate problems of not being able to find the dependency file, the
original dependency file is bundled up with the destination fat object and
when used is unbundled and passed to the aoc compilation.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>