[SYCL][Driver] Emit a warning when appending to an existing archive (#1250)
When llvm-ar is invoked by the driver and the designated output
archive already exists, the behavior is to simply append the new
objects. As such, this GCC-consistent behavior can be quite
useful: e.g. in our SYCL FPGA flow, the user could look to store
a host object file in an archive, then produce an AOT-compiled
device object and append it to the archive via -fsycl-link.

However, if the archive already contains an AOT-compiled device
object, appending a new device object would imminently crash
the further compilation stages. Therefore, a warning is now
provided in case the existing device object-containing archive
under the same name was accindentally left over by the developer.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>