[SYCL] Add buffer dimensions restriction (#1147)
According to the specification, a buffer can have dimensions of only 1, 2, and 3. In the old version, "static_assert" was used to limit it, which did not throw an exception when creating a pointer to a buffer with an incorrect dimension. In this patch it is impossible to create not only an object of the buffer class with incorrect dimensions, but also a link to it.
Also, constructor calls in the accessor class that tried to instantiate a buffer with a zero dimension were fixed.

Signed-off-by: amochalo <anastasiya.mochalova@intel.com>