[SYCL][CUDA] LIT cleanup exceptions
Do not catch unexpected exceptions to return an error code but allow
the exception to terminate the tested program. This will result in
useful exception-related error output that is otherwise lost.

Do not catch and then ignore an unexpected exception but re-throw it.

Explicitly returning error codes.

Abort when catching asynchronous exceptions.

Signed-off-by: Bjoern Knafla <bjoern@codeplay.com>