[SYCL][FPGA] Emit multiple IR variants for the IVDep attribute
By emitting the legacy variant of the LLVM IR alongside the newer
representation of the attribute, backwards compatibility with any
existing BE implementation is restored. A smooth transition period
is thus achieved for the aforementiond BE - until it's able to consume
the new LLVM IR, it has an option to simply ignore the unknown metadata.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>