[SYCL][Driver] Enforce unique filenames when -save-temps is used
With the introduction of file table transformations, multiple
intermediate files with the same extension are now occurring
within any SYCL compilation (namely .txt and .table files).
Unless we treat this during the static creation of comands,
filename collision is bound to occur when saving temps.
This breaks -save-temps compilations due to files being
overwritten.

This commit ensures that unique files are created upon a
potential collision in the presence of -save-temps.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>