[SYCL] Only export public API (#1456)
To enable better control over ABI compatibility, only export functions
that users are expected to use. A public API is the part of SYCL Runtime
that is implemented in the dynamic Runtime library and is expected to be
accessed by the user. That is all public SYCL API (as defined by the
spec), implemented in *.cpp files and all routines accessed by these
APIs (also implemented in *.cpp).

Signed-off-by: Alexander Batashev <alexander.batashev@intel.com>