[Driver][SYCL] Move include/sycl header before other system header locations
The location in which the include/sycl headers are added are too late in
the ordering sequence.  This is causing the possibility of the wrong version
of the OpenCL headers to be picked up (i.e. from /usr/include)

Improve the ordering by putting the dirs much earlier in the command line
instead of relying on the ToolChain addition.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>