[SYCL] Move type checks to later in Semantic Analysis lifecycle (#1465)
SPIR-V doesn't support certain types (quad type, __int128, zero length arrays and other). So we check any code that is running on the kernel for those types and emit an error if they are encountered.

Presently, these checks are done in `ConvertDeclSpecToType` in SemaType.cpp, but this is too early in the Sema lifecycle. It can catch only the most straightforward type declarations. We need to also catch types that use `auto` declarations, `typedef`, templating and more.  To do that, the type checking must occur a little later in the Semantic Analysis lifecycle.

Here I am calling the checks from `CheckCompleteVariableDeclaration` in SemaDecl.cpp. This is called just after the parsing is finished and seems to work well. Also, these checks now avoid the problems we encountered with deferred diagnostics getting confused by templated functions. So, no further change is needed to the deferred diagnostic system (as far as these checks are concerned).

Expanded the testing as well. 

Signed-off-by: Chris Perkins <chris.perkins@intel.com>