[SYCL][PI][CUDA] Implements get_native interoperability (#1332)
* [SYCL][PI][CUDA] Implements get_native interoperability

Implements get_native for CUDA allowing queries for native handles on
SYCL objects; queue, event, context, and device.

* [SYCL][PI][OpenCL] Generalizing interop handler getters

This commit makes get_mem and get_queue of interop_handler return
types based on a specified backend. The backend defaults to OpenCL to
avoid breakages.

* [SYCL][PI] Splitting the converter function
* [SYCL][PI] PI objects created with native handles takes ownership

PI objects created with the associated piextCreate*WithNativeHandle
function takes ownership of the native handle passed to it. For OpenCL
this means that the native OpenCL object is not retained, so the PI
object effectively owns the reference of the caller.
The OpenCL interoperability have been changed to retain the native
OpenCL handle as the implementation no longer does it. This is required
by the SYCL 1.2.1 specification (Rev6, section 4.3.1.)

Signed-off-by: Steffen Larsen <steffen.larsen@codeplay.com>