[Driver][SYCL][FPGA] Improve expected triple and link behaviors for AOCX files
When performing -fsycl-link=image behaviors on Windows, there was a mismatch
of triples being used when checking for existence of the AOCX binary.  This
caused a problem when using clang-cl as the triples do not properly match.
We also do not want to do any additional unbundling of any AOCX archive, which
was causing a conflict of expected FPGA archives.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>