[Driver][SYCL] Remove COFF object format designator for Windows device compiles
When compiling with clang with the Windows target or with clang-cl, the
device triple used was different.  There seems to be no reason for this
difference and only causes mismatches and confusion when using the
different drivers even though they are both targeting Windows.  Effectively
remove this format designator when working with the device triple.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>