Replace local clone Part 1
This patch removes the local clone generated in kernel. In the kernel
body, it replaces any usage of the clone with that of the kernel
object passed as a parameter to the kernel.

Since we are now directly using the kernel object, scalar arguments
(except pointers) and struct type arguments (except those containing
accessor fields) need not be generated and passed to kernel. This
code has been removed.

Pointers require additional 'handling' for USM. For these, arguments
are still generated and passed to kernel. Initialization of clone
pointer fields have been replaced by assignment of kernel object
pointer fields from these parameters.

Please note accessors and streams have not been handled yet. So related
tests will fail as required initializations are now missing.

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>