[SYCL][CUDA] Remove CUDA OpenCL from selection
The NVIDIA OpenCL platform is problematic for NVIDIA CUDA backend users
and for overall DPCPP users, since it doesnt work straight away and is
typically selected on OpenCL backend as a preference.

This patch removes the CUDA OpenCL platform from the device selection,
and prevents it from being used in the lit testing.

Signed-off-by: Ruyman Reyes <ruyman@codeplay.com>