[SYCL] do proactive fix for USM implementation
The idea of the fix is to guarantee that the temporary buffer created
for USM memory does not write to that USM memory at the buffer descturtor.

The write to USM buffer happens explicitly using handler::copy(TmpAcc,USMPtr).
Another write to USMPtr inside temporary buffer destructor could/might happen
even after USMPtr is deallocated, which would cause Seg Fault, which is fixed
in this patch.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>