[SYCL] Add the ability to install the OpenCL ICD loader with LLVM
In certain cases one will want to install the same OpenCL ICD library
along with LLVM that the code was built against. The new option,
OpenCL_INSTALL_KHRONOS_ICD_LOADER can be used to set this up.

- Add the (conditional) installation of the OpenCL ICD library to the
deploy-sycl-toolchain target.
- Add a protection for not trying to install the OpenCL ICD library, if
it is not being built. This is meant to avoid running into issues when
the user mis-configures the build.