[SYCL] Add usm_device and usm_host address spaces
This patch introduces 2 new address spaces in OpenCL: usm_device
and usm_host which are a subset of a global address space.

We want to give the user a way to tell the compiler the allocation
type of a USM pointer for optimization purposes. While it is usually
easy for our compiler to distinguish loads or stores that access local
memory from those that access global memory, distinguishing USM pointers
that access host memory from those that access device memory or even
distinguishing USM pointers that access host memory from accessors that
access global memory is currently impossible. This is because all
host code has been stripped out before we reach the backend and both
accessors and USM pointers are presenting in LLVM IR as pointers in the
global address space in the kernel's arguments. Being able to distinguish
between these types of pointers at compile time is valuable because it allows
us to instantiate simpler load-store units to perform memory transactions.

Signed-off-by: Dmitry Sidorov <dmitry.sidorov@intel.com>