[SYCL] Use a global flush buffer in stream
Change stream implementation to use a global buffer for all the
work-items to flush their outputs into. Previously a set of local flush
buffers were used for this purpose.

This change is made to avoid the issue of some devices running out of
local memory when available space is less than statement size times the
number of work-items.

Signed-off-by: Shahab Layeghi <shahab.layeghi@intel.com>