[SYCL] Add external arbitrary floating point functions
Add support of functions that operate on arbitrary precision floating
point numbers. This datatype and its corresponding operations can be
useful on targets that can take advantage of narrower representation
such as FPGAs.

Arbitrary precision floating point numbers are represented using
arbitrary precision integers (_ExtInt(N)). Additional parameters are:
* _E_: represents the number of exponent bits
* _M_: represents the number of mantissa bits
The total width of the type is `E+M+1` where the last bit is used to
represent the sign. The data layout is shown below:
`[ S (sign bit) ][ E (Exponent) ][ M (Mantissa) ]`

The width of the data `(E+M+1)` is encoded with the width of the
arbitrary precision integer variable. Mantissa (namely `M`) is encoded
within each of the signature of the operations. Exponent values
(namely `E`) are inferred from the width of corresponding variable.

Usage example:
  _ExtInt(Wa) a;
  _ExtInt(Wb) b;
  _ExtInt(Wout) res = __spirv_ArbitraryFloatAddINTEL<Wa, Wb, Wout>(a, Ma, b, Mb, Mout);

Signed-off-by: Mikhail Lychkov <mikhail.lychkov@intel.com>