[SYCL][CUDA] Fix unexpected async memcpy
When memory buffers are created with either of the following flags
* PI_MEM_FLAGS_HOST_PTR_USE
* PI_MEM_FLAGS_HOST_PTR_COPY
we copy the data to the device.

During memory buffer creation we do not have a CUDA stream and therefore call
cuMemCpyHtoD which operates on the CUDA default stream.

This fix synchronizes with the default stream to ensure that data copying is
finished before any other PI operation uses it on a non-default stream.

Signed-off-by: Bjoern Knafla <bjoern@codeplay.com>