[Driver][SYCL] Fix processing when using -fsycl-link (#1765)
When using -fsycl-link, the expectation is for the generated object file
to be able to be linked in by a non-SYCL enabled compiler/linker.  The
generated object did not contain the proper device binary.  Fix this up
to go through all of the proper steps to create the wrapped .spv.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>