[SYCL] Introduce the Level Zero plugin (#1718)
Plugin itself consists of the header and the source file plus cmake
file to build the plugin. Also the following changes were made to
suport the Level Zero plugin in SYCL RT:
* New level0 value was added to the backend enum
* New PI_LEVEL0 value support was added to the SYCL_BE config.
* Docs were updated. Mentioned Level Zero backend and provided the
link to the Level Zero runtime for Intel GPU.
* Changes in sycl cmake file to build Level Zero plugin by default and to
install it with sycl toolchain.

LIT testing with PI_LEVEL0 backend will be enabled in the following
commits. This commits introduces the plugin and makes it buildable.

Signed-off-by: Artur Gainullin artur.gainullin@intel.com