[SYCL] Eliminated redundancy in pi initialization.
Passed the implementation of pi initialization into a "call_once" function, so
that plugin initiliazation only needs to be executed once. After the first time,
the already initiliazed plugin will be returned.

Also, updated pi::initialize to return const reference to the global vector of
plugins, rather than a temporary copy of it.