[Driver][SYCL] Improve handling of linker inputs for static lib processing
When processing the command line for static archives, we go through the
direct to linker values (-Wl, -Xlinker) to find any additional libraries
and objects that should be included.  The trigger options for that scanning
was too broad, allowing too many options (including options with separate
arguments) to be incorrectly processed as objects and/or libraries.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>