[SYCL] Eliminate the redundancy in pi initialization. (#1868)
Passed the implementation of pi initialization into a "call_once" function, so
that plugin initialization only needs to be executed once. After the first time,
the already initialized plugin will be returned.

Also, updated pi::initialize to return const reference to the global vector of
plugins, rather than a temporary copy of it.