[Driver][SYCL][FPGA] Do not generate host dependency files by default
When compiling for -fsycl -fintelfpga, dependency file information needs
to be generated and added to the AOT compilation step.  The dep file is
only needed from the device compilation.

Remove the generation of the dependency files for the host compilation
unless the user explicitly calls for it (with a generation option)

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>