Matching kernel's function types with actual types.
Using getFullyQualifiedType to calculcate the kernel function's types makes the
types in the kernel different from the actual type and that creates a type
mismatch.
getFullyQualifiedType's use is restricted to make type lookup correct in
templates.
This change, replacing the fully qualified type with the canonical
one, fixes the issue, but will imply that the pretty-printer needs to be
updated to print fully qualified names.