[SYCL][CUDA] Fix events for enqueue commands
This commit ensures that `piEnqueueMemBufferCopy`,
`piEnqueueMemBufferFill` and other commands call
both `start()` and `record()` correctly if events are
requested from them. An assertion is added to ensure
this always happens from now on.

Some command types are also changed to reflect the
actual command, for example `cuda_piEnqueueMemBufferReadRect`
not using the `RECT` version of the buffer read
command type.

Some event parameter names are also changed to provide
naming consistency throughout the code.

Signed-off-by: Przemek Malon <przemek.malon@codeplay.com>