[SYCL][CUDA] Lower OpenCL image type to CUDA image descriptors (#1978)
In CUDA images are descriptors represented as an 64 bits integer. The PR lowers the OpenCL image type to this descriptor so it can be processed properly by the NVPTX backend and allow image builtins implementation for this target.

Independent image/sampler are not supported by CUDA, they will be emulated in the libclc. To allow this emulation we need to lower the sampler type to i32. The sampled image type (return type of __spirv_SampledImage) is lowered into a struct containing the image descriptor (i64) and the sampler (i32).