[NFC][SYCL] Move visitors into their own class. (#2081)
A recent patch has shown there is good reason to have access to
Sema/ASTContext when doing the visiting.  This also makes it more
consistent with the other visitors in Clang.

Additionally, there was an odd bracket issue that I've fixed, seemingly
a close/open namespace got lost along the way, so the 'close' namespace
(with comment) bracket was ACTUALLY the end of the function.