[SYCL][CUDA] Enable multiple CUDA arch in the same SYCL build.
The patch refactored the SYCL offloading to account for bound arch as well as toolchains.
Compilation target are represented as a pair of a toolchain (one toolchain per target triple) and a bound arch (which encodes the target arch).

The LTO step is now performed per toolchain/bound arch pair and
the produced binary are then bundled per toolchain.

The patch also makes the bundler/unbundler use the bound arch when SYCL offloading if enabled.

Signed-off-by: Victor Lomuller <victor@codeplay.com>
Co-authored-by: Alexander Johnston <alexander@codeplay.com>