[SYCL] Move deferred emit check to after "Can be emitted" checks.
The previous implementation caused function declarations to be attempted
to be emitted if they were required to be emitted (such as in the case
of -femit-all-decls), which caused a crash.

This patch moves the check until after the 'can be emitted' checks for
function/var decls, and limits it to only functions that are
sycl-devices (or, in the case of implicit attributes, called from a
sycl-device function).