[SYCL] Allow recursive function calls in a constexpr context. (#2105)
It doesn't really make sense to restrict recursion for something that is
forced to be evaluated at constexpr time, since this is a restriction
due to device limitations.  This patch creates a constexpr-context count
for all of the constexpr contexts I could create, and creates a counter.

It is implemented this way to permit a future implementer to add other
diagnostics (such as DLLImport?) as permissible in constexpr.