[SYCL][CUDA] Fix 3D buffer D2D copy slice pitch (#2046)
This commit fixes the slice pitch parameters
for the 2D/3D buffer device to device
copy operation. This fixes a LIT test
(`buffer_full_copy`) that reported passing on
CUDA despite throwing a CUDA error. This LIT test
is amended to check the validity of 3D buffer
copy operations.

The CUDA error occured due to the `srcHeight`
member of the CUDA_MEMCPY3D struct not reflecting
the true extent of the source volume, but the copied
region instead.

Signed-off-by: Przemek Malon <przemek.malon@codeplay.com>