[SYCL] Allow recursive function calls on sycl device with a warning
Currently recursive function calls are not allowed on device.
A hard error is generated in the FE when recursive calls are detected
on device.  To permit enabling this functionality in the BE, this
diagnostic has been moved to the -Wsycl-strict group.  When
-Wsycl-strict is specified, this is still an error, but can be turned
into a warning with -Wno-error=sycl-strict.

TO DO: Once full support is available in the BE, this new diagnostic,
the existing enumerated diagnostic for recursive calls and the
mechanism for detecting the recursion can be removed.

Signed-off-by: Premanand M Rao <premanand.m.rao@intel.com>