[SYCL] Fix undefined behaviour during graph cleanup
With the switch to using the command member field to mark nodes as
visited during graph traversal, deleting nodes during the traversal
leads to invalid memory access whenever there's a node with multiple
indirect dependencies on another node. This patch fixes the issue by
moving the deletion of the nodes to take place post-traversal.