[SYCL] Enable user-defined sub-group reductions (#2106)
Restricted to TriviallyCopyable types for now.

Signed-off-by: John Pennycook <john.pennycook@intel.com>

---

This is a partial implementation of the functionality proposed by @psteinbrecher in https://github.com/intel/llvm/issues/1947, enabling custom sub-group reductions for trivially copyable types.  I cannot see any good way to implement user-defined reductions using `group` instead of `sub_group` today, or to extend support for arbitrary types, but this is a step in the right direction.

Enabling the other algorithms for user-defined types and operators (with the same trivially copyable restriction) could follow the same approach.  Starting with reductions in this PR because they are the most common request.