[SYCL] Disallow mutable lambdas (#1785)
[SYCL] Disallow mutable lambdas (#1785) 

With the latest changes in the SYCL spec kernel lambdas (and functors) must
be const references. Have updated the SYCL headers and tests with this
change. Also had to make a minor adjustment to the kernel generation code
in Clang and update many of the SemaSYCL tests, because their mock kernel
defs needed to be updated as well. Clang also emits a diagnostic if version
 flag (-std-sycl) doesn't match kernel passing (new by const reference or old
 by value) and enabled -std-sycl=2020  
 
Signed-off-by: Chris Perkins chris.perkins@intel.com